--- app/src/main/java/com/example/playlistmaker/compose/AppState.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.compose
-
-data class AppState(
-    val isDarkTheme: Boolean
-)
