+++ app/src/main/java/com/example/playlistmaker/compose/BottomNavigationBar.kt
@@ -37,15 +37,13 @@ fun BottomNavBar(navController: NavHostController) {
                 label = { Text(text = item.name) },
                 selected = currentRoute == item.route,
                 onClick = {
-//                    if (currentRoute != item.route) {
-                        navController.navigate(item.route) {
-                            popUpTo(navController.graph.findStartDestination().id) {
-                                saveState = true
-                            }
-                            launchSingleTop = true
-                            restoreState = true
+                    navController.navigate(item.route) {
+                        popUpTo(navController.graph.findStartDestination().id) {
+                            saveState = true
                         }
-//                    }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
                 },
                 colors = NavigationBarItemDefaults.colors(
                     selectedIconColor = MaterialTheme.colorScheme.primary,
