+++ app/src/main/java/com/example/playlistmaker/main/navigation/NavGraph.kt
@@ -60,7 +60,7 @@ fun NavGraph(
                 )
             }
             composable(route = NavRoutes.NewPlaylistPage.route) {
-                NewPlaylistPage()
+                NewPlaylistPage(onBackClick = { navController.popBackStack() })
             }
         }
     }
